% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDT.R, R/generatewidgets.R
\name{generate_widget.DT}
\alias{generate_widget.DT}
\alias{generate_widget}
\alias{generate_widget.default}
\title{Generate widget}
\usage{
\method{generate_widget}{DT}(x, ...)

generate_widget(x, ...)

\method{generate_widget}{default}(x, ...)
}
\arguments{
\item{x}{unpacked widget meta data}

\item{...}{additional arguments to be passed in generating individual widgets}
}
\value{
\code{NULL}

htmlwidget object

\code{NULL}
}
\description{
S3 generic method of generating a widget
}
\examples{
## DT examples
library(magrittr)
opts <- list(scrollY = 300, scroller = TRUE)
list(package = 'DT', type = 'table', data = 'iris') \%>\%
    unpack_data() \%>\% generate_widget(update_size = TRUE)
list(package = 'DT', type = 'table', data = 'iris', options = opts) \%>\%
    unpack_data() \%>\% generate_widget(update_size = TRUE)
}
