% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_widget.R
\name{unpack_data}
\alias{unpack_data}
\title{Unpack data}
\usage{
unpack_data(x)
}
\arguments{
\item{x}{widget meta data}
}
\value{
widget meta data given class
}
\description{
\code{unpack_data} assigns class to widget meta data
}
