% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWidgets.R
\name{update_dep_path}
\alias{update_dep_path}
\title{Update dependency path}
\usage{
update_dep_path(dep, libdir = "lib")
}
\arguments{
\item{dep}{dep}

\item{libdir}{libdir}
}
\value{
dep
}
\description{
\code{update_dep_path} updates dependency path
}
\examples{
\dontrun{

}
}
