% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatewidgets.R
\name{update_sizing_policy}
\alias{update_sizing_policy}
\title{Update sizing policy}
\usage{
update_sizing_policy(x)
}
\arguments{
\item{x}{widget meta data}
}
\value{
widget meta data given class
}
\description{
\code{update_sizing_policy} updates browser sizing policy of a widget
}
