% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_widget.R
\name{write_widget}
\alias{write_widget}
\title{Write widget}
\usage{
write_widget(widget, cdn = NULL, output_option = NULL,
  background = "white")
}
\arguments{
\item{widget}{widget}

\item{cdn}{cdn}

\item{output_option}{option}

\item{background}{background}
}
\value{
widget
}
\description{
\code{write_widget} writes a widget
}
\examples{
library(magrittr)
dt <- list(package = 'DT', type = 'table', data = 'iris') \%>\%
    unpack_data() \%>\% generate_widget(update_size = TRUE)
write_widget(widget = dt, cdn = NULL, output_option = 'html')
}
