% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWidgets.R
\name{write_widget}
\alias{write_widget}
\title{Write widget}
\usage{
write_widget(widget, cdn = NULL, output_option = NULL,
  background = "white")
}
\arguments{
\item{widget}{widget}

\item{cdn}{cdn}

\item{output_option}{option}

\item{background}{background}
}
\value{
widget
}
\description{
\code{write_widget} write widgets
}
\examples{
\dontrun{

}
}
