% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWidgets.R
\name{write_widget}
\alias{write_widget}
\title{Write widget}
\usage{
write_widget(widget, libdir = "lib", cdn = NULL, background = "white",
  to_save = FALSE, to_raw = FALSE)
}
\arguments{
\item{widget}{widget}

\item{libdir}{libdir}

\item{cdn}{cdn}

\item{background}{background}

\item{to_save}{to_save}

\item{to_raw}{to_raw}
}
\value{
widget
}
\description{
\code{write_widget} write widgets
}
\examples{
\dontrun{

}
}
